feat: unified float dtype system with float32 default
Add package-wide dtype configuration backed by contextvars.ContextVar:
- mdpp.set_default_dtype() / get_default_dtype()
- mdpp.default_dtype() context manager for scoped override
- Per-function dtype= parameter (highest precedence)
- Default is np.float32 (matches MD simulation precision)

Every compute/featurize function now accepts dtype= parameter.
Functions requiring float64 for numerical stability (DCCM covariance,
FES log-probability, PCA/TICA via sklearn/deeptime) compute internally
in float64 and cast output to the resolved dtype.

Dataclass fields changed from NDArray[np.float64] to NDArray[np.floating].
Resolution: function dtype > context manager > set_default_dtype > float32.
419 tests pass across the full suite.

+++ tests/analysis/test_delta_rmsf.py
@@ -102,7 +102,7 @@ class TestComputeDeltaRMSFIdentical:
         a1 = _make_rmsf([0.1, 0.3])
         a2 = _make_rmsf([0.3, 0.1])
         b = _make_rmsf([0.2, 0.2])
-        result = compute_delta_rmsf([a1, a2], [b])
+        result = compute_delta_rmsf([a1, a2], [b], dtype=np.float64)
 
         avg_a = np.sqrt(np.mean([np.array([0.1, 0.3]) ** 2, np.array([0.3, 0.1]) ** 2], axis=0))
         expected = np.array([0.2, 0.2]) - avg_a
